moved underscore wrapper and added tapi18n project file to build
